getting rid of abstractions, introducing 'transactional' flag for InMemoryEventStoreFactory
